Split resource/row tests
